pyproject.toml
[project]
name = "ai-healthassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.0",
    "bitsandbytes>=0.47.0",
    "fastapi>=0.116.1",
    "gtts>=2.5.4",
    "hf-xet>=1.1.8",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.6",
    "openai-whisper>=20250625",
    "peft>=0.17.1",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "speechrecognition>=3.14.3",
    "streamlit>=1.48.1",
    "transformers>=4.55.4",
    "uvicorn>=0.35.0",
]
